requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
anyio==3.5.0
APScheduler==3.8.1
asgiref==3.5.0
async-timeout==3.0.1
asyncmy==0.2.3
atomicwrites==1.4.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
discord==1.7.3
discord-webhook==0.14.0
discord.py==1.7.3
fastapi==0.73.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
networkx==2.8.3
numpy==1.22.2
packaging==21.3
pandas==1.4.0
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3339==6.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.3
XlsxWriter==3.0.2
yarl==1.7.2
